refactor: add thin application entrypoint
