+++ cmd/puremania/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"os"
+
+	"puremania/internal/app"
+)
+
+func main() {
+	os.Exit(app.Run(os.Args[1:]))
+}
